requirements.txt
PyQt5>=5.15.0
PyQtWebEngine>=5.15.0
opencv-python>=4.5.0
numpy>=1.19.0
requests>=2.25.0
Pillow>=8.0.0
scikit-learn>=0.24.0
flask>=2.0.0
paddleocr>=2.0.0
streamlit
beautifulsoup4
selenium
webdriver-manager
pywebview>=4.0
numpy
opencv-python
DrissionPage>=4.0.0